Finalizado a parte da implementação funcional do total de aparições em filmes por planeta.

+++ src/main/java/com/b2w/starwar/domain/entity/Planeta.java
@@ -30,6 +30,6 @@ public class Planeta {
 
     private String terreno;
 
-    private Long totalDeAparicoesEmFilmes;
+    private Integer totalDeAparicoesEmFilmes;
 
 }
